requirements.txt
scikit-learn==1.3.2
numpy==1.24.4
streamlit
pillow
numpy
